medguard_backend/requirements.txt
# Core Django and Wagtail
# Note: Django 7.0.2 doesn't exist yet. Latest stable version is 5.2.4
Django==5.2.4
wagtail==7.0.2

# Database
psycopg2-binary==2.9.10

# Wagtail dependencies
django-modelcluster==6.4
django-taggit==6.1.0
django-treebeard==4.7.1
django-permissionedforms==0.1
draftjs_exporter==5.1.0
telepath==0.3.1
Willow==1.11.0

# Image processing
Pillow==11.3.0
pillow_heif==1.0.0

# Environment and configuration
python-dotenv==1.0.0
django-environ==0.12.0

# API and serialization
djangorestframework==3.16.0
django-cors-headers==4.7.0

# Filtering and search
django-filter==25.1

# Background tasks
celery==5.5.3
redis==6.2.0

# Utilities
python-dateutil==2.9.0.post0
requests==2.32.4
beautifulsoup4==4.13.4

# Development and testing
django-stubs-ext==5.2.2
django-debug-toolbar==4.4.0

# Wagtail admin enhancements
wagtail-modeladmin==2.2.0

# File handling
openpyxl==3.1.5
filetype==1.2.0

# Security
defusedxml==0.7.1

# Internationalization
anyascii==0.3.3
